Add a Docstring
I added this Docstring so that I can talk about the columns of the table
in the database. There are 7 columns, but the user will only see 5. The
remaining 2 will never be seen by the user but we will use those two
columns to perform very important tasks. The tasks that we will be
performing with those two columns(PriortityNumber and StatusNumber) is
that they will help us to sort the data in the database by using the
status or priority of the data and also help display the data based on
priority or status or even delete all data in the database that have
been fulfilled or have a past date.