Add another column to database and create update_status function
A new column have been added to the database called DaysLeftOrOverdue,
it purpose is to store the number of days left for the day for the
list's status to turn 'in progress', overdue or completed from pending.
It will have zero if list is ongoing.

About the update_status function, it updates all the status of the data
in the database and updates the number of days left for the data
to turn ongoing or overdue. So it basic purpose is to update the data
in the database.