Apply .__str__() to the answer of the subtracted dates before splitting
When .__str__() is applied to a subtracted date variable, the outcome
looks like '60 days, 0:00:00', and it is only the first character
after the first space which is a number and in this case 60, is the only
information that we need to work with. So I have applied that so that
the program won't run into errors.