Add add_new_item() function
The new add_new_item() function collects all data needed from the user
to create a row in the database table for a to-do item.

There are functions that were called in the add_new_item() to get a
due date from the user or use the due data to perform some calculations
and return some values that will be needed in some of the table fields.
The functions are due_date(), days_left(), item_status() and
item_status_number(). days_left(), item_status() and
item_status_number() take arguments to perform their calculations.