Create get_due_date() function then create get_month() function
get_due_date() gets date, calls get_month() to get the month from the
user and then get the year from the user. It accepts years that are
greater or equal to current year. It returns this information to
add new_item() function as the date entered by the user.