Create items_status(), days_left() and item_status_number()
items_status() function takes the number of days left as an argument and
returns the status based on the number of days left. If the number of
days left is 0, it returns 'in progress' and if the number of days left
is greater than 0, it returns 'pending'.

days_left() function takes the date entered by the user as an argument
and returns the number of number days left for the day to be due.

item_status_number() produces a number that will be used by the
application when sorting output and the function get status as an
argument and returns this unique number.