Update item_status() and item_status_number()
I thought I was going to update get_due_date() while I was writing the
code for item_status() function and item_status_number() function. I
was going to update it to take only days and months that have not passed
like I did for the year, but looking at the whole code, I will have to
restructure the code to be able to do that. I don't want to go through
that pain, so I just left it like that. If that needs done, I will
take care of it in the future, but now I think this should work well...