Edit functions called by view_an_item() & edit update_a_list_item()
I added some code to view_by_desc(), view_by_priority() and
view_by_status() so that they can be used by other functions.

I added some code to update_a_list_item(), some if else statements and
a logic for when a user want to update a description. I have called
update_description() function and passed the ID number of the particular
item that the user wants to change it description to it. This is done
after the user finish view the item or an item in the to-do list.