Write the logic for update_a_list_item() function
Elif statement that get triggered when the user choose to update
priority have been coded. It calls view_an_item() function and for
the user to be able to view an item in the to-do_list before updating
it priority status. It calls update_priority() function and passes in
an argument which is the id number of the item which it priority needs
to get changed.