Add delete_item() function and get_description() function
delete_item() ask the user if they want to delete a particular item or
all items, base on the user selection delete_an_item() function which
is also created will be called or delete_all_items() function will be
called.

delete_an_item() function help the user to delete just an item and
delete_all_items() get the user to delete all items.

get_description() function get's a get_description that is less than
25 characters from the user. It have been used in the function that
adds new items to the to do list and also the function that updates
the description of a to do list item.