Add a functionality to the program where an item can be marked completed
First of all, I have made changes in update_status() function, so that
when it get run and there is any item with a status as completed, it
won't change it status to overdue(if the date for completion is passed),
or 'In progress' if the current day is the day it supposed to be done.

Secondly, I have added an option to update_a_list_item() function, so
that a user can mark an item in the to-do list completed. I have also
introduced an elif statement for that to work and in the course of doing
that, I had to add a function that performs the functionality.

The function I added was mark_status_completed().
mark_status_completed() function get an argument from update_a_list_item
function which is the id of the item for it status to be marked
completed. If the item truly exist in the database,
mark_status_completed() function will ask the user if they are sure that
they want to mark the status of the item that they provided it id as
completed. If the user says yes, the status will be marked as completed.