Import calendar module, introduce a function called last_day_of_month()
get_due_date() function have been edited to handle situations where the
user might enter a day greater than the last dy of the month. This led
to the introduction of a new function called last_day_of_month(). This
new function creation called for the importation of calendar module.
What this new function does is to get the last day of the month. This
day is returned to get_due_date() function and it is used to restrict
the user from entering a day that is larger than the number of days in
a month.