Edit main() and view_by_desc()
The main() function was using recursion, because we don't want issues
at run time, we replaced the recursion with a while loop. While loop
was introduced in view_by_desc() function to take care of some few
stuff. Some comments were added to some parts of the code to help me and
anyone who will come and read in the future to get a better
understanding of what is going on.